Put all the chord parsing stuff in a whole nother file.
